Create class ResultsTable and move most of UIStarter into it
The differences between the new ResultsTable and the old UIStarter are:
- ResultsTable doesn't have an init() or launch() method, but adds a constructor and a ready() method
- Results table has a UIStarter instance variable, which it receives in its constructor
- ResultsTable replaces its uiCleanup() method with a quit() method, which calls uiCleanup in UIStarter
- ResultsTable doesn't have setupIcons() or run()

Now FileCopyMonitor and ProgressBarUpdater deal with a ResultsTable, not a UIStarter.

UIStarter still sets up the Display, the Shell, the icons, etc.  But everything specific to the table is moved into ResultsTable.
